Address review: encode LAG/LEAD offset as int4
PostgreSQL's lag/lead take an `integer` (int4) offset, but the offset was stored
and rendered as i64 (BIGINT), causing a parameter-type mismatch on prepare/
execute. Store and bind it as i32 (as ntile already does); lag/lead now take an
i32 offset. Updated the Postgres test to expect Int32.

+++ crates/squealy-postgresql/tests/sql.rs
@@ -2312,7 +2312,7 @@ fn postgres_window_functions_render_with_numbered_placeholders() {
     assert!(sql.contains("LAG(q0_0.\"user_id\", $2)"), "{sql}");
     assert_eq!(
         args.collect_params().unwrap(),
-        vec![PostgresParam::Int32(4), PostgresParam::Int64(2)]
+        vec![PostgresParam::Int32(4), PostgresParam::Int32(2)]
     );
 
     // Aggregate-over wraps the cast around the whole window expression, not the inner aggregate.
